used ordinary Integer key type

+++ src/main/java/space/bumtiger/repository/CorderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import space.bumtiger.domain.Corder;
 
 public interface CorderRepository 
-				extends CrudRepository<Corder, Long> {
+				extends CrudRepository<Corder, Integer> {
 }
